requirements.txt
langchain
langchain-core
langchain-community 
pinecone-client
Sentence-Transformers
PyPDF2
FastAPI
pydantic
gradio
groq